refactor: partially modularize
