+++ src/main.rs
@@ -4,24 +4,25 @@ extern crate pest;
 #[macro_use]
 extern crate pest_derive;
 
-use pest::error::Error;
-use pest::Parser;
 use core::panic;
 use std::collections::{HashMap, BTreeMap};
 use std::env;
-use std::fs::{canonicalize, read, read_to_string, File};
+use std::fs::{canonicalize, read_to_string, File};
 use std::fmt::Debug;
 use std::path::PathBuf;
 use std::process::exit;
-use std::rc::Rc;
-use std::cell::{Cell, RefCell};
 use std::sync::Mutex;
 use std::io::Write;
-use std::ops::Deref;
 
-#[derive(Parser)]
-#[grammar = "fox32.pest"]
-struct Fox32Parser;
+pub (crate) mod instr;
+pub(crate) mod parser;
+pub(crate) mod util;
+pub(crate) mod include;
+
+use instr::{instruction_to_byte, AssembledInstruction, Condition, InstructionIncDec, InstructionTwo, OperationIncDec, OperationOne, OperationTwo, OperationZero, Size};
+use parser::{AstNode, backpatch::{BackpatchTarget, perform_backpatching}};
+use util::format_address_table;
+use include::include_text_file;
 
 // this is kinda dumb, but oh well !!
 lazy_static! {
@@ -38,310 +39,12 @@ lazy_static! {
 const FXF_RELOC_SIZE:  usize = 0x0000000C;
 const FXF_RELOC_PTR:   usize = 0x00000010;
 
-#[derive(Debug, Clone)]
-struct BackpatchTarget {
-    index: usize,
-    size: Size,
-    is_relative: bool,
-    instruction: AssembledInstruction,
-}
-
-impl BackpatchTarget {
-    fn new(instruction: &AssembledInstruction, index: usize, size: Size, is_relative: bool) -> BackpatchTarget {
-        Self {
-            index, is_relative, size,
-            instruction: instruction.clone(),
-        }
-    }
-
-    fn write(&self, size: Size, address: u32) {
-        let ref instruction = self.instruction;
-        let mut instruction_data = instruction.borrow_mut();
-
-        let address_bytes =
-            if self.is_relative {
-                (address as i32 - self.instruction.get_address() as i32).to_le_bytes()
-            } else {
-                address.to_le_bytes()
-            };
-
-        match size {
-            Size::Byte => instruction_data[self.index] = address_bytes[0],
-            Size::Half => {
-                instruction_data[self.index]     = address_bytes[0];
-                instruction_data[self.index + 1] = address_bytes[1];
-            },
-            Size::Word => {
-                instruction_data[self.index]     = address_bytes[0];
-                instruction_data[self.index + 1] = address_bytes[1];
-                instruction_data[self.index + 2] = address_bytes[2];
-                instruction_data[self.index + 3] = address_bytes[3];
-            }
-        }
-    }
-
-    fn get_backpatch_location(&self) -> u32 {
-        self.instruction.get_address() + self.index as u32
-    }
-}
-
-fn perform_backpatching(targets: &Vec<BackpatchTarget>, address: (u32, bool)) {
-    for target in targets {
-        target.write(target.size, address.0);
-
-        // if this label isn't const or relative, then add it to the reloc table for FXF
-        if !address.1 && !target.is_relative {
-            let mut reloc_table = RELOC_ADDRESSES.lock().unwrap();
-            reloc_table.push(target.get_backpatch_location());
-        }
-    }
-}
-
-#[derive(Debug, Clone, Default)]
-struct AssembledInstruction {
-    value: Rc<RefCell<Vec<u8>>>,
-    address: Rc<Cell<u32>>,
-}
-
-impl AssembledInstruction {
-    fn new() -> Self {
-        Self {
-            value: Rc::default(),
-            address: Rc::default(),
-        }
-    }
-
-    fn get_address(&self) -> u32 {
-        self.address.get()
-    }
-    fn set_address(&self, address: u32) {
-        self.address.set(address);
-    }
-}
-
-impl From<Vec<u8>> for AssembledInstruction {
-    fn from(data: Vec<u8>) -> Self {
-        Self {
-            value: Rc::new(RefCell::new(data)),
-            address: Rc::default(),
-        }
-    }
-}
-
-impl From<&[u8]> for AssembledInstruction {
-    fn from(data: &[u8]) -> Self {
-        Vec::from(data).into()
-    }
-}
-
-impl<const N: usize> From<[u8; N]> for AssembledInstruction {
-    fn from(data: [u8; N]) -> Self {
-        (&data[..]).into()
-    }
-}
-
-impl Deref for AssembledInstruction {
-    type Target = RefCell<Vec<u8>>;
-
-    fn deref(&self) -> &Self::Target {
-        &self.value
-    }
-}
-
-unsafe impl Send for AssembledInstruction {}
-unsafe impl Sync for AssembledInstruction {}
-
-#[derive(PartialEq, Debug, Clone, Copy)]
-enum InstructionZero {
-    // no operands
-    Nop,
-    Halt,
-    Brk,
-    Ret,
-    Reti,
-    Ise,
-    Icl,
-    Mse,
-    Mcl,
-}
-
-#[derive(PartialEq, Debug, Clone, Copy)]
-enum InstructionOne {
-    // one operand
-    Not,
-    Jmp,
-    Call,
-    Loop,
-    Rjmp,
-    Rcall,
-    Rloop,
-    Push,
-    Pop,
-    Int,
-    Tlb,
-    Flp,
-}
-
-#[derive(PartialEq, Debug, Clone, Copy)]
-enum InstructionIncDec {
-    // one or two operands
-    Inc,
-    Dec,
-}
-
-#[derive(PartialEq, Debug, Clone, Copy)]
-enum InstructionTwo {
-    // two operands
-    Add,
-    Sub,
-    Mul,
-    Imul,
-    Div,
-    Idiv,
-    Rem,
-    Irem,
-    And,
-    Or,
-    Xor,
-    Sla,
-    Sra,
-    Srl,
-    Rol,
-    Ror,
-    Bse,
-    Bcl,
-    Bts,
-    Cmp,
-    Mov,
-    Movz,
-    Rta,
-    In,
-    Out,
-}
-
-#[derive(PartialEq, Debug, Clone, Copy)]
-enum Size {
-    Byte,
-    Half,
-    Word,
-}
-
-#[derive(PartialEq, Debug, Clone, Copy)]
-enum Condition {
-    Always,
-    Zero,
-    NotZero,
-    Carry,
-    NotCarry,
-    GreaterThan,
-    // GreaterThanEqualTo is equivalent to NotCarry
-    // LessThan is equivalent to Carry
-    LessThanEqualTo,
-}
-
-#[derive(PartialEq, Debug, Clone, Copy)]
-enum LabelKind {
-    Internal,
-    External,
-    Global,
-}
-
-#[derive(PartialEq, Debug, Clone)]
-struct OperationZero {
-    size: Size,
-    condition: Condition,
-    instruction: InstructionZero,
-}
-#[derive(PartialEq, Debug, Clone)]
-struct OperationOne {
-    size: Size,
-    condition: Condition,
-    instruction: InstructionOne,
-    operand: Box<AstNode>,
-}
-#[derive(PartialEq, Debug, Clone)]
-struct OperationIncDec {
-    size: Size,
-    condition: Condition,
-    instruction: InstructionIncDec,
-    lhs: Box<AstNode>,
-    rhs: Box<AstNode>,
-}
-#[derive(PartialEq, Debug, Clone)]
-struct OperationTwo {
-    size: Size,
-    condition: Condition,
-    instruction: InstructionTwo,
-    lhs: Box<AstNode>,
-    rhs: Box<AstNode>,
-}
-
 // Used by data.fill to store either a known size or the name of a constant
 // that defines the size
 #[derive(PartialEq, Debug, Clone)]
 enum SizeOrLabelName {
     Size(u32),
     Label(String),
-}
-
-#[derive(PartialEq, Debug, Clone)]
-enum AstNode {
-    OperationZero(OperationZero) ,
-    OperationOne (OperationOne),
-    OperationIncDec(OperationIncDec) ,
-    OperationTwo (OperationTwo),
-
-    Immediate8(u8),
-    Immediate16(u16),
-    Immediate32(u32),
-    Register(u8),
-    ImmediatePointer(u32),
-    RegisterPointer(u8),
-    RegisterPointerOffset(u8, u8),
-
-    Constant {
-        name: String,
-        address: u32,
-    },
-
-    LabelDefine {
-        name: String,
-        kind: LabelKind,
-    },
-    LabelOperand {
-        name: String,
-        size: Size,
-        is_relative: bool,
-    },
-    LabelOperandPointer {
-        name: String,
-        is_relative: bool,
-    },
-
-    DataByte(u8),
-    DataHalf(u16),
-    DataWord(u32),
-    DataStr(String),
-    DataStrZero(String),
-    DataFill {
-        value: u8,
-        size: SizeOrLabelName,
-    },
-
-    IncludedBinary(Vec<u8>),
-
-    Origin(u32),
-    OriginPadded(u32),
-    Optimize(bool)
-}
-
-fn format_address_table(m: &HashMap<String, (u32, bool)>) -> String {
-    let mut v: Vec<(&String, &u32)> = Vec::new();
-    for i in m.into_iter() {
-        v.push((i.0, &i.1.0));
-    }
-    v.sort_by(|(_, v1), (_, v2)| u32::cmp(v1, v2));
-    v.iter().map(|(k, v)| format!("{:#010X?} :: {}", v, k)).collect::<Vec<String>>().join("\n")
 }
 
 fn main() {
@@ -383,7 +86,7 @@ fn main() {
     }
 
     println!("Parsing file...");
-    let ast = match parse(&input_file) {
+    let ast = match parser::parse(&input_file) {
         Ok(x) => x,
         Err(x) => {
             println!("{:#?}", x);
@@ -519,602 +222,6 @@ fn main() {
     output_file.write_all(&binary).unwrap();
 }
 
-fn include_text_file(line_number: usize, text: &str, input_file: String) -> String {
-    //println!("{}, {}", line_number, text);
-    let path_start_index = text.find("\"").unwrap() + 1;
-    let path_end_index = text.len() - 1;
-    let path_string = &text[path_start_index..path_end_index];
-    //let path = canonicalize(path_string).expect(&format!("failed to include file \"{}\"", path_string));
-
-    let mut source_path = SOURCE_PATH.lock().unwrap().clone();
-    source_path.push(path_string);
-
-    println!("Including file as text data: {:#?}", source_path.file_name().expect("invalid filename"));
-
-    let mut start_of_original_file = String::new();
-    for (i, text) in input_file.lines().enumerate() {
-        if i < line_number {
-            start_of_original_file.push_str(text);
-            start_of_original_file.push('\n');
-        }
-    }
-
-    let mut included_file = read_to_string(source_path).expect(&format!("failed to include file \"{}\"", path_string));
-    included_file.push('\n');
-
-    let mut end_of_original_file = String::new();
-    for (i, text) in input_file.lines().enumerate() {
-        if i > line_number {
-            end_of_original_file.push_str(text);
-            end_of_original_file.push('\n');
-        }
-    }
-
-    let mut final_file = String::new();
-
-    final_file.push_str(&start_of_original_file);
-    final_file.push_str(&included_file);
-    final_file.push_str(&end_of_original_file);
-    final_file
-}
-
-fn include_binary_file(pair: pest::iterators::Pair<Rule>, optional: bool) -> AstNode {
-    let path_string = pair.into_inner().next().unwrap().as_str().trim();
-
-    let mut source_path = SOURCE_PATH.lock().unwrap().clone();
-    source_path.push(path_string);
-
-    println!("Including file as binary data: {:#?}", source_path.file_name().expect("invalid filename"));
-
-    let binary = read(&source_path);
-    if binary.is_err() && optional {
-        println!("Optional include was not found: {:#?}", source_path.file_name().expect("invalid filename"));
-        return AstNode::IncludedBinary(vec![]);
-    } else if binary.is_err() {
-        panic!("failed to include file");
-    }
-
-    AstNode::IncludedBinary(binary.unwrap())
-}
-
-fn parse(source: &str) -> Result<Vec<AstNode>, Error<Rule>> {
-    let mut ast = vec![];
-    let pairs = Fox32Parser::parse(Rule::assembly, source)?;
-
-    for pair in pairs.peek().unwrap().into_inner() {
-        match pair.as_rule() {
-            Rule::EOI => break,
-            _ => ast.push(build_ast_from_expression(pair)),
-        }
-    }
-
-    Ok(ast)
-}
-
-fn build_ast_from_expression(pair: pest::iterators::Pair<Rule>) -> AstNode {
-    //println!("{:#?}\n\n", pair); // debug
-    let pair_rule = pair.as_rule();
-    let mut inner_pair = pair.into_inner();
-    *CURRENT_CONDITION.lock().unwrap() = Condition::Always;
-    let mut is_pointer = false;
-    match inner_pair.peek().unwrap().as_rule() {
-        Rule::condition => {
-            *CURRENT_CONDITION.lock().unwrap() = parse_condition(&inner_pair.peek().unwrap());
-            inner_pair.next().unwrap(); // jump to the next instruction pair after the condition
-        }
-        Rule::operand_value_ptr => {
-            is_pointer = true;
-        }
-        _ => {}
-    }
-
-    match pair_rule {
-        Rule::assembly => build_ast_from_expression(inner_pair.next().unwrap()),
-        Rule::instruction => parse_instruction(inner_pair.next().unwrap()),
-        Rule::operand => parse_operand(inner_pair.next().unwrap(), is_pointer),
-        Rule::constant => parse_constant(inner_pair),
-        Rule::label => parse_label(inner_pair.next().unwrap(), inner_pair.next()),
-        Rule::data => parse_data(inner_pair.next().unwrap()),
-        Rule::opt => parse_opt(inner_pair.next().unwrap()),
-        Rule::origin => parse_origin(inner_pair.next().unwrap()),
-        Rule::include_bin => include_binary_file(inner_pair.next().unwrap(), false),
-        Rule::include_bin_optional => include_binary_file(inner_pair.next().unwrap(), true),
-        _ => todo!("{:#?}", pair_rule),
-    }
-}
-
-fn parse_constant(pairs: pest::iterators::Pairs<Rule>) -> AstNode {
-    *CURRENT_SIZE.lock().unwrap() = Size::Word;
-    let mut pairs = pairs;
-    let constant_name = pairs.next().unwrap().into_inner().next().unwrap().as_str();
-    let operand_pair = pairs.next().unwrap();
-    let operand_ast = parse_operand(operand_pair, false);
-
-    if let AstNode::Immediate32(address) = operand_ast {
-        AstNode::Constant {
-            name: constant_name.to_string(),
-            address,
-        }
-    } else {
-        panic!("Constant must be an immediate value");
-    }
-}
-
-fn parse_label(pair: pest::iterators::Pair<Rule>, next_pair: Option<pest::iterators::Pair<Rule>>) -> AstNode {
-    let mut name_pair = pair.clone();
-    let kind = match pair.as_rule() {
-        Rule::label_kind => {
-            let pair_inner = pair.clone().into_inner().next().unwrap();
-            name_pair = next_pair.unwrap();
-            match pair_inner.as_rule() {
-                Rule::label_external => LabelKind::External,
-                Rule::label_global => LabelKind::Global,
-                _ => unreachable!()
-            }
-        },
-        _ => LabelKind::Internal,
-    };
-    let node = AstNode::LabelDefine {name: name_pair.as_str().to_string(), kind};
-    node
-}
-
-fn parse_data(pair: pest::iterators::Pair<Rule>) -> AstNode {
-    //println!("{:#?}", pair);
-    *CURRENT_SIZE.lock().unwrap() = Size::Word;
-    match pair.as_rule() {
-        Rule::data_byte => {
-            match parse_operand(pair.into_inner().next().unwrap(), false) {
-                AstNode::Immediate32(half) => AstNode::DataByte(half as u8),
-                AstNode::LabelOperand {name, size: _, is_relative} =>
-                    AstNode::LabelOperand {name, size: Size::Byte, is_relative},
-                _ => unreachable!(),
-            }
-        },
-        Rule::data_half => {
-            match parse_operand(pair.into_inner().next().unwrap(), false) {
-                AstNode::Immediate32(half) => AstNode::DataHalf(half as u16),
-                AstNode::LabelOperand {name, size: _, is_relative} =>
-                    AstNode::LabelOperand {name, size: Size::Half, is_relative},
-                _ => unreachable!(),
-            }
-        },
-        Rule::data_word => {
-            match parse_operand(pair.into_inner().next().unwrap(), false) {
-                AstNode::Immediate32(word) => AstNode::DataWord(word),
-                AstNode::LabelOperand {name, size: _, is_relative} =>
-                    AstNode::LabelOperand {name, size: Size::Word, is_relative},
-                _ => unreachable!(),
-            }
-        },
-        Rule::data_str => {
-            let string = pair.into_inner().next().unwrap().into_inner().next().unwrap().as_str();
-            AstNode::DataStr(string.to_string())
-        },
-        Rule::data_strz => {
-            let string = pair.into_inner().next().unwrap().into_inner().next().unwrap().as_str();
-            AstNode::DataStrZero(string.to_string())
-        },
-        Rule::data_fill => {
-            let value = {
-                let ast = parse_operand(pair.clone().into_inner().next().unwrap(), false);
-                if let AstNode::Immediate32(word) = ast {
-                    word as u8
-                } else {
-                    unreachable!()
-                }
-            };
-            let size = {
-                let ast = parse_operand(pair.into_inner().nth(1).unwrap(), false);
-                if let AstNode::Immediate32(word) = ast {
-                    SizeOrLabelName::Size(word)
-                } else if let AstNode::LabelOperand {name, ..} = ast {
-                    SizeOrLabelName::Label(name)
-                } else {
-                    dbg!(ast);
-                    unreachable!()
-                }
-            };
-            AstNode::DataFill {value, size}
-        },
-        _ => panic!("Unsupported data: {}", pair.as_str()),
-    }
-}
-fn parse_opt(rule: pest::iterators::Pair<Rule>) -> AstNode {
-    match rule.as_str() {
-        "opton"=>AstNode::Optimize(true),
-        "optoff"=>AstNode::Optimize(false),
-        _ => panic!("Unknown optimize flag {}", rule.as_str())
-    }
-}
-fn parse_origin(pair: pest::iterators::Pair<Rule>) -> AstNode {
-    //println!("{:#?}", pair);
-    match pair.as_rule() {
-        Rule::origin_no_padding => {
-            let ast = parse_operand(pair.into_inner().next().unwrap(), false);
-            let address = {
-                if let AstNode::Immediate32(word) = ast {
-                    word
-                } else {
-                    unreachable!()
-                }
-            };
-            AstNode::Origin(address)
-        },
-        Rule::origin_padding => {
-            let ast = parse_operand(pair.into_inner().next().unwrap(), false);
-            let address = {
-                if let AstNode::Immediate32(word) = ast {
-                    word
-                } else {
-                    unreachable!()
-                }
-            };
-            AstNode::OriginPadded(address)
-        },
-        _ => panic!("Unsupported origin: {}", pair.as_str()),
-    }
-}
-
-fn parse_size(pair: &pest::iterators::Pair<Rule>) -> Size {
-    match pair.as_str() {
-        ".8" => Size::Byte,
-        ".16" => Size::Half,
-        ".32" => Size::Word,
-        _ => panic!("Unsupported size: {}", pair.as_str()),
-    }
-}
-
-fn parse_incdec_amount(pair: pest::iterators::Pair<Rule>) -> AstNode {
-    match pair.as_str() {
-        "1" => AstNode::Immediate8(0),
-        "2" => AstNode::Immediate8(1),
-        "4" => AstNode::Immediate8(2),
-        "8" => AstNode::Immediate8(3),
-        _ => panic!("Unsupported increment/decrement: {}", pair.as_str()),
-    }
-}
-
-fn parse_condition(pair: &pest::iterators::Pair<Rule>) -> Condition {
-    match pair.as_str() {
-        "ifz" => Condition::Zero,
-        "ifnz" => Condition::NotZero,
-        "ifc" => Condition::Carry,
-        "ifnc" => Condition::NotCarry,
-        "ifgt" => Condition::GreaterThan,
-        "ifgteq" => Condition::NotCarry,
-        "iflt" => Condition::Carry,
-        "iflteq" => Condition::LessThanEqualTo,
-        _ => panic!("Unsupported condition: {}", pair.as_str()),
-    }
-}
-
-fn parse_instruction(pair: pest::iterators::Pair<Rule>) -> AstNode {
-    //println!("parse_instruction: {:#?}", pair); // debug
-    let mut size = Size::Word;
-    let condition = *CURRENT_CONDITION.lock().unwrap();
-    match pair.as_rule() {
-        Rule::instruction_conditional => {
-            let mut inner_pair = pair.into_inner();
-            let instruction_conditional_pair = inner_pair.next().unwrap();
-            match instruction_conditional_pair.as_rule() {
-                Rule::instruction_zero => {
-                    if let Some(inner) = inner_pair.peek() {
-                        if inner.as_rule() == Rule::size {
-                            size = parse_size(&inner_pair.next().unwrap());
-                        }
-                    }
-                    *CURRENT_SIZE.lock().unwrap() = size;
-                    parse_instruction_zero(instruction_conditional_pair, size, condition)
-                }
-                Rule::instruction_one => {
-                    if inner_pair.peek().unwrap().as_rule() == Rule::size {
-                        size = parse_size(&inner_pair.next().unwrap());
-                    }
-                    *CURRENT_SIZE.lock().unwrap() = size;
-                    let operand = inner_pair.next().unwrap();
-                    let operand_ast = build_ast_from_expression(operand);
-                    parse_instruction_one(instruction_conditional_pair, operand_ast, size, condition)
-                }
-                Rule::instruction_incdec => {
-                    if inner_pair.peek().unwrap().as_rule() == Rule::size {
-                        size = parse_size(&inner_pair.next().unwrap());
-                    }
-                    *CURRENT_SIZE.lock().unwrap() = size;
-                    let lhs = inner_pair.next().unwrap();
-                    let lhs_ast = build_ast_from_expression(lhs);
-                    let rhs_ast = if inner_pair.peek().is_some() {
-                        let rhs = inner_pair.next().unwrap();
-                        parse_incdec_amount(rhs)
-                    } else {
-                        AstNode::Immediate8(0)
-                    };
-                    parse_instruction_incdec(instruction_conditional_pair, lhs_ast, rhs_ast, size, condition)
-                }
-                Rule::instruction_two => {
-                    if inner_pair.peek().unwrap().as_rule() == Rule::size {
-                        size = parse_size(&inner_pair.next().unwrap());
-                    }
-                    *CURRENT_SIZE.lock().unwrap() = size;
-                    let lhs = inner_pair.next().unwrap();
-                    let rhs = inner_pair.next().unwrap();
-                    let lhs_ast = build_ast_from_expression(lhs);
-                    let rhs_ast = build_ast_from_expression(rhs);
-                    parse_instruction_two(instruction_conditional_pair, lhs_ast, rhs_ast, size, condition)
-                }
-                _ => todo!(),
-            }
-        }
-        _ => panic!("Unsupported instruction type: {:#?}", pair.as_rule()),
-    }
-}
-
-fn remove_underscores(input: &str) -> String {
-    String::from_iter(input.chars().filter(|c| *c != '_'))
-}
-
-fn immediate_to_astnode(immediate: u32, size: Size, is_pointer: bool) -> AstNode {
-    if is_pointer {
-        AstNode::ImmediatePointer(immediate)
-    } else {
-        match size {
-            Size::Byte => AstNode::Immediate8(immediate as u8),
-            Size::Half => AstNode::Immediate16(immediate as u16),
-            Size::Word => AstNode::Immediate32(immediate),
-        }
-    }
-}
-
-fn parse_immediate(pair: pest::iterators::Pair<Rule>) -> u32 {
-    match pair.as_rule() {
-        Rule::immediate_bin => {
-            let body_bin_str = pair.into_inner().next().unwrap().as_str();
-        u32::from_str_radix(&remove_underscores(body_bin_str), 2).unwrap()
-        }
-        Rule::immediate_hex => {
-            let body_hex_str = pair.into_inner().next().unwrap().as_str();
-            u32::from_str_radix(&remove_underscores(body_hex_str), 16).unwrap()
-        }
-        Rule::immediate_dec => {
-            let dec_str = pair.as_span().as_str();
-            remove_underscores(dec_str).parse::<u32>().unwrap()
-        }
-        Rule::immediate_char => {
-            let body_char_str = pair.into_inner().next().unwrap().as_str();
-            body_char_str.chars().nth(0).unwrap() as u8 as u32
-        }
-        _=> {
-            panic!()
-        }
-    }
-}
-
-fn parse_register(pair: pest::iterators::Pair<Rule>) -> u8 {
-    let register_num_pair = pair.into_inner().next().unwrap();
-    let register_num = if register_num_pair.as_str() == "sp" { 32 }
-    else if register_num_pair.as_str() == "esp" { 33 }
-    else if register_num_pair.as_str() == "fp" { 34 }
-    else { register_num_pair.as_str().parse::<u8>().unwrap() };
-    if register_num > 34 { panic!("register number out of range"); }
-    register_num
-}
-
-fn parse_operand(mut pair: pest::iterators::Pair<Rule>, is_pointer: bool) -> AstNode {
-    //println!("parse_operand: {:#?}", pair); // debug
-    // dbg!(&pair);
-    let size = *CURRENT_SIZE.lock().unwrap();
-    let pointer_offset = 
-    if is_pointer {
-        // skip past the operand_value_ptr pair and look at its operand_value rule
-        let mut pairs = pair.into_inner();
-        pair = pairs.next().unwrap();
-        pairs.next()
-        // pair = pair.into_inner().next().unwrap();
-    }else {
-        None
-    };
-    match pair.as_rule() {
-        Rule::operand_value => {
-            let mut inner_pair = pair.into_inner();
-            let operand_value_pair = inner_pair.next().unwrap();
-            match operand_value_pair.as_rule() {
-                Rule::immediate_bin|
-                Rule::immediate_char|
-                Rule::immediate_dec|
-                Rule::immediate_hex => {
-                    immediate_to_astnode(parse_immediate(operand_value_pair), size, is_pointer)
-                }
-                Rule::register => {
-                    let register_num = parse_register(operand_value_pair);
-                    if is_pointer {
-                        AstNode::RegisterPointer(register_num)
-                    } else {
-                        AstNode::Register(register_num)
-                    }
-                }
-                Rule::label_name => {
-                    if is_pointer {
-                        AstNode::LabelOperandPointer {
-                            name: operand_value_pair.as_str().to_string(),
-                            is_relative: false,
-                        }
-                    } else {
-                        AstNode::LabelOperand {
-                            name: operand_value_pair.as_str().to_string(),
-                            size,
-                            is_relative: false,
-                        }
-                    }
-                }
-                _ => todo!(),
-            }
-        }
-        Rule::register => {
-            let register_num = parse_register(pair);
-            let offset = if let Some(offset_pair) = pointer_offset {
-                parse_immediate(offset_pair.into_inner().next().unwrap())
-            } else {
-                0
-            };
-            if offset == 0 {
-                AstNode::RegisterPointer(register_num)
-            } else {
-                AstNode::RegisterPointerOffset(register_num, offset as u8)
-            }
-        }
-        _ => panic!(),
-    }
-}
-
-fn parse_instruction_zero(pair: pest::iterators::Pair<Rule>, size: Size, condition: Condition) -> AstNode {
-    AstNode::OperationZero ( OperationZero {
-        size: size,
-        condition: condition,
-        instruction: match pair.as_str() {
-            "nop"  => InstructionZero::Nop,
-            "halt" => InstructionZero::Halt,
-            "brk"  => InstructionZero::Brk,
-            "ret"  => InstructionZero::Ret,
-            "reti" => InstructionZero::Reti,
-            "ise"  => InstructionZero::Ise,
-            "icl"  => InstructionZero::Icl,
-            "mse"  => InstructionZero::Mse,
-            "mcl"  => InstructionZero::Mcl,
-            _ => panic!("Unsupported conditional instruction (zero): {}", pair.as_str()),
-        }
-    })
-}
-
-fn parse_instruction_one(pair: pest::iterators::Pair<Rule>, mut operand: AstNode, size: Size, condition: Condition) -> AstNode {
-    AstNode::OperationOne ( OperationOne {
-        size: size,
-        condition: condition,
-        instruction: match pair.as_str() {
-            "not"   => InstructionOne::Not,
-            "jmp"   => InstructionOne::Jmp,
-            "call"  => InstructionOne::Call,
-            "loop"  => InstructionOne::Loop,
-            "rjmp"  => {
-                match &mut operand {
-                    &mut AstNode::LabelOperand        {ref mut is_relative, ..} |
-                    &mut AstNode::LabelOperandPointer {ref mut is_relative, ..} => {
-                        *is_relative = true;
-                    }
-                    _ => {}
-                }
-                InstructionOne::Rjmp
-            },
-            "rcall" => {
-                match &mut operand {
-                    &mut AstNode::LabelOperand        {ref mut is_relative, ..} |
-                    &mut AstNode::LabelOperandPointer {ref mut is_relative, ..} => {
-                        *is_relative = true;
-                    }
-                    _ => {}
-                }
-                InstructionOne::Rcall
-            },
-            "rloop" => {
-                match &mut operand {
-                    &mut AstNode::LabelOperand        {ref mut is_relative, ..} |
-                    &mut AstNode::LabelOperandPointer {ref mut is_relative, ..} => {
-                        *is_relative = true;
-                    }
-                    _ => {}
-                }
-                InstructionOne::Rloop
-            },
-            "push"  => InstructionOne::Push,
-            "pop"   => InstructionOne::Pop,
-            "int"   => InstructionOne::Int,
-            "tlb"   => InstructionOne::Tlb,
-            "flp"   => InstructionOne::Flp,
-            _ => panic!("Unsupported conditional instruction (one): {}", pair.as_str()),
-        },
-        operand: Box::new(operand)
-    })
-}
-
-fn parse_instruction_incdec(pair: pest::iterators::Pair<Rule>, lhs: AstNode, rhs: AstNode, size: Size, condition: Condition) -> AstNode {
-    AstNode::OperationIncDec ( OperationIncDec {
-        size: size,
-        condition: condition,
-        instruction: match pair.as_str() {
-            "inc"  => InstructionIncDec::Inc,
-            "dec"  => InstructionIncDec::Dec,
-            _ => panic!("Unsupported conditional instruction (two): {}", pair.as_str()),
-        },
-        lhs: Box::new(lhs),
-        rhs: Box::new(rhs),
-    })
-}
-
-
-fn parse_instruction_two(pair: pest::iterators::Pair<Rule>, mut lhs: AstNode, mut rhs: AstNode, size: Size, condition: Condition) -> AstNode {
-    match pair.as_str() {
-        "sla"  |
-        "sra"  |
-        "srl"  |
-        "rol"  |
-        "ror"  |
-        "bse"  |
-        "bcl"  |
-        "bts"  => if let Some(value) = node_value(&rhs) {
-            rhs = AstNode::Immediate8(value as u8);
-        }
-        _=>()
-    }
-    AstNode::OperationTwo ( OperationTwo {
-        size: size,
-        condition: condition,
-        instruction: match pair.as_str() {
-            "add"  => InstructionTwo::Add,
-            "sub"  => InstructionTwo::Sub,
-            "mul"  => InstructionTwo::Mul,
-            "imul" => InstructionTwo::Imul,
-            "div"  => InstructionTwo::Div,
-            "idiv" => InstructionTwo::Idiv,
-            "rem"  => InstructionTwo::Rem,
-            "irem" => InstructionTwo::Irem,
-            "and"  => InstructionTwo::And,
-            "or"   => InstructionTwo::Or,
-            "xor"  => InstructionTwo::Xor,
-            "sla"  => InstructionTwo::Sla,
-            "sra"  => InstructionTwo::Sra,
-            "srl"  => InstructionTwo::Srl,
-            "rol"  => InstructionTwo::Rol,
-            "ror"  => InstructionTwo::Ror,
-            "bse"  => InstructionTwo::Bse,
-            "bcl"  => InstructionTwo::Bcl,
-            "bts"  => InstructionTwo::Bts,
-            "cmp"  => InstructionTwo::Cmp,
-            "mov"  => InstructionTwo::Mov,
-            "movz" => InstructionTwo::Movz,
-            "rta"  => {
-                match &mut lhs {
-                    &mut AstNode::LabelOperand        {ref mut is_relative, ..} |
-                    &mut AstNode::LabelOperandPointer {ref mut is_relative, ..} => {
-                        *is_relative = true;
-                    }
-                    _ => {}
-                }
-                match &mut rhs {
-                    &mut AstNode::LabelOperand        {ref mut is_relative, ..} |
-                    &mut AstNode::LabelOperandPointer {ref mut is_relative, ..} => {
-                        *is_relative = true;
-                    }
-                    _ => {}
-                }
-                InstructionTwo::Rta
-            }
-            "in"   => InstructionTwo::In,
-            "out"  => InstructionTwo::Out,
-            _ => panic!("Unsupported conditional instruction (two): {}", pair.as_str()),
-        },
-        lhs: Box::new(lhs),
-        rhs: Box::new(rhs),
-    })
-}
-
 fn assemble_node(node: AstNode) -> AssembledInstruction {
     // if this is data, don't interpret it as an instruction
     match node {
@@ -1168,84 +275,6 @@ fn assemble_node(node: AstNode) -> AssembledInstruction {
 //     node_to_immediate_values(&node, &mut vec);
 //     vec
 // }
-
-fn size_to_byte(size: Size) -> u8 {
-    match size {
-        Size::Byte => 0b00000000,
-        Size::Half => 0b01000000,
-        Size::Word => 0b10000000,
-    }
-}
-
-fn instruction_to_byte(node: &AstNode) -> u8 {
-    match *node {
-        AstNode::OperationZero (OperationZero{size, instruction, ..}) => {
-            match instruction {
-                InstructionZero::Nop  => 0x00 | size_to_byte(size),
-                InstructionZero::Halt => 0x10 | size_to_byte(size),
-                InstructionZero::Brk  => 0x20 | size_to_byte(size),
-                InstructionZero::Ret  => 0x2A | size_to_byte(size),
-                InstructionZero::Reti => 0x3A | size_to_byte(size),
-                InstructionZero::Ise  => 0x0C | size_to_byte(size),
-                InstructionZero::Icl  => 0x1C | size_to_byte(size),
-                InstructionZero::Mse  => 0x0D | size_to_byte(size),
-                InstructionZero::Mcl  => 0x1D | size_to_byte(size),
-            }
-        }
-        AstNode::OperationOne (OperationOne{size, instruction, ..}) => {
-            match instruction {
-                InstructionOne::Not   => 0x33 | size_to_byte(size),
-                InstructionOne::Jmp   => 0x08 | size_to_byte(size),
-                InstructionOne::Call  => 0x18 | size_to_byte(size),
-                InstructionOne::Loop  => 0x28 | size_to_byte(size),
-                InstructionOne::Rjmp  => 0x09 | size_to_byte(size),
-                InstructionOne::Rcall => 0x19 | size_to_byte(size),
-                InstructionOne::Rloop => 0x29 | size_to_byte(size),
-                InstructionOne::Push  => 0x0A | size_to_byte(size),
-                InstructionOne::Pop   => 0x1A | size_to_byte(size),
-                InstructionOne::Int   => 0x2C | size_to_byte(size),
-                InstructionOne::Tlb   => 0x2D | size_to_byte(size),
-                InstructionOne::Flp   => 0x3D | size_to_byte(size),
-            }
-        }
-        AstNode::OperationIncDec (OperationIncDec{size, instruction, ..}) => {
-            match instruction {
-                InstructionIncDec::Inc   => 0x11 | size_to_byte(size),
-                InstructionIncDec::Dec   => 0x31 | size_to_byte(size),
-            }
-        }
-        AstNode::OperationTwo (OperationTwo{size, instruction, ..}) => {
-            match instruction {
-                InstructionTwo::Add  => 0x01 | size_to_byte(size),
-                InstructionTwo::Sub  => 0x21 | size_to_byte(size),
-                InstructionTwo::Mul  => 0x02 | size_to_byte(size),
-                InstructionTwo::Imul => 0x14 | size_to_byte(size),
-                InstructionTwo::Div  => 0x22 | size_to_byte(size),
-                InstructionTwo::Idiv => 0x34 | size_to_byte(size),
-                InstructionTwo::Rem  => 0x32 | size_to_byte(size),
-                InstructionTwo::Irem => 0x35 | size_to_byte(size),
-                InstructionTwo::And  => 0x03 | size_to_byte(size),
-                InstructionTwo::Or   => 0x13 | size_to_byte(size),
-                InstructionTwo::Xor  => 0x23 | size_to_byte(size),
-                InstructionTwo::Sla  => 0x04 | size_to_byte(size),
-                InstructionTwo::Sra  => 0x05 | size_to_byte(size),
-                InstructionTwo::Srl  => 0x15 | size_to_byte(size),
-                InstructionTwo::Rol  => 0x24 | size_to_byte(size),
-                InstructionTwo::Ror  => 0x25 | size_to_byte(size),
-                InstructionTwo::Bse  => 0x06 | size_to_byte(size),
-                InstructionTwo::Bcl  => 0x16 | size_to_byte(size),
-                InstructionTwo::Bts  => 0x26 | size_to_byte(size),
-                InstructionTwo::Cmp  => 0x07 | size_to_byte(size),
-                InstructionTwo::Mov  => 0x17 | size_to_byte(size),
-                InstructionTwo::Movz => 0x27 | size_to_byte(size),
-                InstructionTwo::Rta  => 0x39 | size_to_byte(size),
-                InstructionTwo::In   => 0x0B | size_to_byte(size),
-                InstructionTwo::Out  => 0x1B | size_to_byte(size),
-            }
-        }
-        _ => panic!("Attempting to parse a non-instruction AST node as an instruction: {:#?}", node),
-    }
-}
 
 fn condition_to_bits(condition: &Condition) -> u8 {
     match condition {
